Tweaks for DSL integration
