+++ Sources/App/Controllers/PackageController.swift
@@ -1,11 +1,14 @@
 import Fluent
+import Plot
 import Vapor
+
 
 struct PackageController {
 
-    func index(req: Request) throws -> EventLoopFuture<[Package]> {
+    func index(req: Request) throws -> EventLoopFuture<HTML> {
         return Package.query(on: req.db)
             .all()
+            .map(SPIPage.packageIndex)
     }
 
     func get(req: Request) throws -> EventLoopFuture<Package> {
