+++ Sources/App/Views/PackageController/PackageControllerIndex.swift
@@ -1,9 +1,14 @@
 import Plot
 
-class PackageControllerIndex: PublicPage {
 
-    override func content() -> Node<HTML.BodyContext> {
-        .main(.p("Hello, World."))
+extension SPIPage {
+    static func packageIndex(_ packages: [Package]) -> HTML {
+        SPIPage.document(
+            .main(
+                .group(
+                    packages.map(\.url).map { .pre(.text($0)) }
+                )
+            )
+        )
     }
-
 }
