+++ Sources/App/routes.swift
@@ -2,13 +2,13 @@ import Fluent
 import Vapor
 
 func routes(_ app: Application) throws {
-    app.get { req in
-        homePage()
-    }
-
     let packageController = PackageController()
     app.get("packages", use: packageController.index)
     app.get("packages", ":id", use: packageController.get)
+
+    app.group("admin") { admin in
+        admin.get { req in SPIPage.admin() }
+    }
 
     app.group("api") { api in
         let apiPackageController = API.PackageController()
